+++ menger-app/src/main/scala/menger/engines/VideoEngine.scala
@@ -8,6 +8,7 @@ import menger.common.ProfilingConfig
 import menger.common.RenderConfig
 import menger.config.ExecutionConfig
 import menger.config.TAnimationConfig
+import menger.dsl.DenoiseMode
 import menger.dsl.Scene
 
 class VideoEngine(
@@ -18,7 +19,8 @@ class VideoEngine(
   val causticsConfig: CausticsConfig,
   val videoOutputPath: String,
   val videoQuality: Int,
-  val keepFrames: Boolean
+  val keepFrames: Boolean,
+  denoiseModeOverride: Option[DenoiseMode] = None
 )(using ProfilingConfig)
     extends BaseEngine(executionConfig.maxInstances)
     with WithAnimation with WithVideoExport with SavesScreenshots:
@@ -31,6 +33,11 @@ class VideoEngine(
 
   override protected val firstFrameConfigs: SceneConverter.SceneConfigs =
     SceneConverter.convert(_firstScene, causticsConfig)
+
+  override protected def denoiseMode: DenoiseMode =
+    denoiseModeOverride.getOrElse(firstFrameConfigs.denoiseMode)
+
+  override protected def accumulationFrames: Int = firstFrameConfigs.accumulationFrames
 
   override protected val sceneConfigurator: SceneConfigurator = SceneConfigurator(
     firstFrameConfigs.camera.position.toVector3,